@@ -1,6 +1,6 @@
+   
 
-    <!-- Vendor Scripts Start -->
-    <script src="{{ asset('assets') }}/js/vendor/jquery-3.5.1.min.js"></script>
+   <!-- Vendor Scripts Start -->
     <script src="{{ asset('assets') }}/js/vendor/bootstrap.bundle.min.js"></script>
     <script src="{{ asset('assets') }}/js/vendor/OverlayScrollbars.min.js"></script>
     <script src="{{ asset('assets') }}/js/vendor/autoComplete.min.js"></script>
@@ -9,19 +9,29 @@
     <script src="{{ asset('assets') }}/icon/acorn-icons.js"></script>
     <script src="{{ asset('assets') }}/icon/acorn-icons-interface.js"></script>
 
-    <script src=" {{ asset('assets') }}/js/vendor/Chart.bundle.min.js"></script>
+    <script src="{{ asset('assets') }}/js/cs/scrollspy.js"></script>
 
-    <script src=" {{ asset('assets') }}/js/vendor/chartjs-plugin-datalabels.js"></script>
+    <script src="{{ asset('assets') }}/js/vendor/jquery.validate/jquery.validate.min.js"></script>
 
-    <script src=" {{ asset('assets') }}/js/vendor/chartjs-plugin-rounded-bar.min.js"></script>
+    <script src="{{ asset('assets') }}/js/vendor/jquery.validate/additional-methods.min.js"></script>
 
-    <script src=" {{ asset('assets') }}/js/vendor/glide.min.js"></script>
+    <script src="{{ asset('assets') }}/js/vendor/datepicker/bootstrap-datepicker.min.js"></script>
 
-    <script src=" {{ asset('assets') }}/js/vendor/intro.min.js"></script>
+    <script src="{{ asset('assets') }}/js/vendor/tagify.min.js"></script>
 
-    <script src=" {{ asset('assets') }}/js/vendor/select2.full.min.js"></script>
+    <script src="{{ asset('assets') }}/js/vendor/Chart.bundle.min.js"></script>
 
-    <script src=" {{ asset('assets') }}/js/vendor/plyr.min.js"></script>
+    <script src="{{ asset('assets') }}/js/vendor/chartjs-plugin-datalabels.js"></script>
+
+    <script src="{{ asset('assets') }}/js/vendor/chartjs-plugin-rounded-bar.min.js"></script>
+
+    <script src="{{ asset('assets') }}/js/vendor/glide.min.js"></script>
+
+    <script src="{{ asset('assets') }}/js/vendor/intro.min.js"></script>
+
+    <script src="{{ asset('assets') }}/js/vendor/select2.full.min.js"></script>
+
+    <script src="{{ asset('assets') }}/js/vendor/plyr.min.js"></script>
 
     <!-- Vendor Scripts End -->
 
@@ -34,12 +44,21 @@
     <!-- Template Base Scripts End -->
     <!-- Page Specific Scripts Start -->
 
-    <script src=" {{ asset('assets') }}/js/cs/glide.custom.js"></script>
+    <script src="{{ asset('assets') }}/js/cs/glide.custom.js"></script>
 
-    <script src=" {{ asset('assets') }}/js/cs/charts.extend.js"></script>
+    <script src="{{ asset('assets') }}/js/cs/charts.extend.js"></script>
 
-    <script src=" {{ asset('assets') }}/js/pages/dashboard.default.js"></script>
+    <script src="{{ asset('assets') }}/js/cs/datatable.extend.js"></script>
+
+    <script src="{{ asset('assets') }}/js/pages/dashboard.default.js"></script>
+
+    <script src="{{ asset('assets') }}/js/forms/validation.js"></script>
 
     <script src="{{ asset('assets') }}/js/common.js"></script>
     <script src="{{ asset('assets') }}/js/scripts.js"></script>
+
     <!-- Page Specific Scripts End -->
+     <!-- Sweet Alert Js -->
+<script src="{{ asset('assets') }}/js/vendor/sweetalert2/sweetalert2.min.js"></script>
+<!-- Toaster Alert Js -->
+<script src="{{ asset('assets') }}/js/vendor/toastr/toastr.min.js"></script>
